Unfucking my fuck up
--> There were files throughout my repo which I didn't want to commit.
As all those files were not finalized and were not upto the mark with
the quality of repo but just to persist them while switching laptop
I didn't git clone this repo from remote as I wanted those uncommited
files also and instead converted the local folder into zip and unzipped
it in a new laptop, and I noticed something new, now git recognizes all
the files as modified due to zipping and unzipping and that's why even
if the content is not changed, still the git marks all the content of
all files as modified.

--> So, now I am going to commit whatever there is in this repo, fuck
it, IDGAF now.

--> Although one new thing which I realised it that this is the only
repo which didn't had any uncommited important files, WOWWW!! so,
just because of other repos this one just suffered in company.

--> So, I can skip this commit and instead git clone this repo at
least directly from remote in this new laptop, but I am lazy, and I
have typed this commit message so I am gonna commit it anyways ;)